Student specific comments
When a Student speaks, you're having to call out to the Chat, then the ChatEvent - the ChatEvent class has significantly added to your overhead for the program, but doesn't really give you a benefit when (in the current model) chat messages are simply Strings. The one thing it added (other than connect/disconnect) is a timestamp, whcih may be worth it in some contexts - that said, I'd definitely implement this piece a little differently to cut down on the back and forth.