ChatWindow feedback
One highly recommended practice in programming (including C and C++ as IDEs became better) is to declare variables as close to their use as possible. This means that if at all possible, you should avoid using "fields" or "instance variables" for elements that are local scope *in usage*. That means even if it's something that belongs to the class as a whole in terms of actual implementation, if I'm only using the variable locally and I don't have to otherwise declare it in the classes scope, I should still only declare it locally.
You have a couple of fields which would most likely fit the category of "better in local scope".